@@ -31,14 +31,13 @@
     "woke": "scripts/cli woke"
   },
   "devDependencies": {
+    "@eslint/js": "9.17.0",
     "@jsdevtools/coverage-istanbul-loader": "3.0.5",
     "@types/chrome": "0.0.287",
     "@types/connect-busboy": "1.0.3",
     "@types/cors": "2.8.17",
     "@types/express": "4.17.21",
     "@types/jasmine": "3.10.18",
-    "@typescript-eslint/eslint-plugin": "8.19.0",
-    "@typescript-eslint/parser": "8.19.0",
     "@wdio/browserstack-service": "8.41.0",
     "@wdio/cli": "8.41.0",
     "@wdio/jasmine-framework": "8.41.0",
@@ -51,17 +50,15 @@
     "connect-busboy": "1.0.0",
     "cors": "2.8.5",
     "emoji-name-map": "1.2.9",
-    "eslint": "8.57.1",
-    "eslint-config-prettier": "9.1.0",
+    "eslint": "9.17.0",
     "eslint-module-utils": "2.12.0",
     "eslint-plugin-import": "2.31.0",
     "eslint-plugin-jasmine": "4.2.2",
     "eslint-plugin-jsdoc": "50.6.1",
-    "eslint-plugin-local-rules": "3.0.2",
-    "eslint-plugin-prefer-arrow": "1.2.3",
     "eslint-plugin-unicorn": "56.0.1",
     "express": "4.21.2",
     "glob": "11.0.0",
+    "globals": "15.12.0",
     "html-webpack-plugin": "5.6.3",
     "jasmine-core": "3.99.1",
     "json-schema-to-typescript": "bcaudan/json-schema-to-typescript#bcaudan/add-readonly-support",
@@ -84,6 +81,7 @@
     "ts-node": "10.9.2",
     "tsconfig-paths-webpack-plugin": "4.2.0",
     "typescript": "5.7.2",
+    "typescript-eslint": "8.19.1",
     "webdriverio": "8.41.0",
     "webpack": "5.97.1",
     "webpack-cli": "6.0.1",
