@@ -1,4 +1,4 @@
-import pako from 'pako'
+import { deflate } from 'pako'
 import type { BrowserSegment, BrowserSegmentMetadata } from '../../types'
 import { readReplayPayload } from '../../../test'
 import { buildReplayPayload } from './buildReplayPayload'
@@ -6,7 +6,7 @@ import { buildReplayPayload } from './buildReplayPayload'
 describe('buildReplayPayload', () => {
   const SEGMENT = { foo: 'bar' } as unknown as BrowserSegment
   const SERIALIZED_SEGMENT = JSON.stringify(SEGMENT)
-  const COMPRESSED_SEGMENT = pako.deflate(SERIALIZED_SEGMENT)
+  const COMPRESSED_SEGMENT = deflate(SERIALIZED_SEGMENT)
   const METADATA: BrowserSegmentMetadata = {
     application: { id: 'xxx' },
     session: { id: 'xxx' },
